mms: also load group message recipients from CC field
If group messaging is enabled by the user, the messaging framework should also
try to load any numbers in the CC PduHeader. Some devices initiate group
messaging by populating this field with the group members' phone numbers, and
they only place the first number in the TO header. This should solve threading
issues with these devices. loadRecipients() will handle the CC field correctly
if it is null.
Signed-off-by: Christian Gaetano <cgatno@gmail.com>

Change-Id: I955bb0912356f4d8fb648acd2bd9e6bba21c9a82